refactor(referrals): one shared page for both dashboards
consultee/referrals (394 LOC) and consultant/referrals (474) were
near-clones. Every /api/referrals/* endpoint keys off the session rather
than a profile id, and the consultant page didn't even read its own
`params` — so the data was already identical and only the presentation
differed.

Extracted to components/dashboard/shared/ReferralsPage, matching the seven
shared operator page components already in that folder. The `role` prop
drives the three real differences: the subtitle, the consultant-only
"Total Earned" card, and the vanity-code input (consultants share the link
publicly; consultees refer people they already know).

868 -> 503 LOC.

Part of #1021
